feat(index.html): add blog link
I've been working on a Tech Blog on&off, where I put in tips, tricks and
sometimes just thoughts, most of which is directly related to what I was
working on that day.
Currently it focuses majorly on Python & Vim, however, I plan on adding
React to it as well.
Further, I updated the about a little bit but not too much, to keep the
original spirit in which the website was written in intact.
Finally, I added Vim(dotfiles repo), React.js and PyTest in the tooling section.